moved union_rdd test to integrated test
